fix: address audit findings — green CI mypy, fail-loud eligibility, vectorize direct
- config.py: drop the unused [import-untyped] ignore on yaml; types-PyYAML is
  in dev deps so the comment was triggering [unused-ignore] in CI strict mode.
- statistical.py: pin dtype on the holm-branch np.arange to unify mypy's
  inferred type for raw_adjusted with the fdr_bh branch.
- pairwise.py: raise DataValidationError instead of silently substituting the
  first operator when (a) a client has no eligible operators (all three
  induction strategies) or (b) a day has clients but no operator candidates
  (stream_topk). Aligns with invariants.md fail-loud preference flagged in
  the prior round of review.
- pairwise.py: replace the O(n_chunk * n_clients) iterrows+scan client-index
  lookup in induce_policy_direct with an O(n_clients) dict build plus an
  O(n_chunk) map; preserves first-occurrence-wins semantics for non-unique
  client_id within a day.
- tests: add parametrized test_empty_eligibility_raises_across_strategies
  covering direct/stream/stream_topk.

Resolves the mypy strict failures that were keeping CI red on the new
Python 3.10-3.14 matrix.

+++ src/skdr_eval/statistical.py
@@ -602,7 +602,7 @@ def multiple_comparison_correction(
         # Holm-Bonferroni step-down method with monotonicity adjustment
         sorted_indices = np.argsort(p_arr)
         sorted_p = p_arr[sorted_indices]
-        raw_adjusted = sorted_p * np.arange(n, 0, -1)
+        raw_adjusted = sorted_p * np.arange(n, 0, -1, dtype=np.float64)
         monotone_adjusted = np.maximum.accumulate(raw_adjusted)
         monotone_adjusted = np.minimum(monotone_adjusted, 1.0)
         corrected = np.zeros_like(p_arr)
